dashboards: Remove grafana-prom in favor of grafana-prom-v2
v2 was initially created because some large changes to the dashboard
were being made, and we wanted to keep the old one for users. It's been
long enough now, and all panels are in v2.

Additionally, both use the same uid. This causes grafana to only load
one of these, seemingly nondeterministically.